hub/requirements.txt
flynn
websockets
pynacl

# windcentrale source
aiohttp
